+++ src/gl-state-glx.h
@@ -44,7 +44,7 @@ public:
     bool init_gl_extensions();
     bool reset();
     void swap();
-    bool gotNativeConfig(int& vid);
+    bool gotNativeConfig(intptr_t& vid);
     void getVisualConfig(GLVisualConfig& vc);
 
 private:
